Copy some enhancements from the async branch
Bailing on the promise of creating a makefile

Fix some implicit narrowing conversions
Remove dependency on boost filesystem in tests
Make clean script also clean libevent

+++ include/AuthenticationRequest.hpp
@@ -33,8 +33,8 @@ public:
     int32_t getSerializedSize() {
         return 8 + //String length prefixes
         20 + //SHA-1 hash of PW
-        m_username.size() +
-        m_service.size()
+        static_cast<int32_t>(m_username.size()) +
+        static_cast<int32_t>(m_service.size())
         + 4 //length prefix
         + 1; //version number
     }
